fix: Resolve PHPStan Level 5 issues in Treasury Portfolio code
- Fixed missing constructor parameters in RebalancingApprovalRequested event
- Added empty array checks before using max() function
- Cast ceil() result to int explicitly
- Fixed test assertions to use correct methods
- Removed invalid method calls in workflow tests
- Fixed ReflectionClass instantiation with proper type checking

These fixes ensure the code passes PHPStan Level 5 static analysis.

+++ app/Domain/Treasury/Activities/Portfolio/GenerateReportActivity.php
@@ -433,7 +433,7 @@ class GenerateReportActivity extends Activity
         // Rough estimation: ~500 words per page
         $wordCount = str_word_count(strip_tags($content));
 
-        return max(1, ceil($wordCount / 500));
+        return (int) max(1, ceil($wordCount / 500));
     }
 
     private function simulatePdfGeneration(string $htmlContent, array $options): string
